W171c: btlFormation FIVE window clears + h_maestro MATCH.
btlFormation over-window 24 -> 19 by replacing invented array zero-fills, u64 masks/widths and untyped pointer/call paths with sibling-grounded typed selection logic, explicit iterator/selected roles, and retail-shaped positive guards and loops. Four of the five are now near-misses: 002CEC30 364/304 nd264 -> 300/304 nd12; 002CF1D0 428/368 nd303 -> 364/368 nd16; 002CEF10 368/352 nd180 -> 352/352 nd18; 002CEB20 328/272 nd234 -> 264/272 nd42; 002CF070 424/352 nd297 -> 340/352 nd178. LF-only endings preserved (0 CRLF, 15,136 LF), 520 markers unchanged, 46/46 pragma brackets, retained opt_common_subs left in place.

h_maestro 001165C0 nd16 -> MATCH by recovering FIVE HIDDEN DESTRUCTOR ARGUMENTS and reordering base/counter declarations. Also 00111150 264/272 nd31 -> nd9 and 00110510 176 -> 180 bytes nd131 -> 113. h_camp untouched; its annotated barriers and confirmed floors left alone.

Combined: over-window bytes 2988 -> 2724, nd -1068, zero lost, zero additions.

+++ src/Graphics/Model/mdlFile.c
@@ -1259,7 +1259,7 @@ void FUN_0034bd60(u32 param_1);
 u32 FUN_0034bdf0(u32 param_1,u32 param_2);
 void FUN_0034bef0();
 u32 FUN_0034bf10(int param_1);
-u64 FUN_0034bfc0(u32 param_1);
+u32 FUN_0034bfc0(u32 param_1);
 void FUN_0034c160(u32 param_1);
 u32 FUN_0034c1b0(int param_1);
 void FUN_0034c250(int param_1,int param_2);
@@ -1402,7 +1402,7 @@ void FUN_00357730(int param_1);
 u32 FUN_00357770(u32 param_1,u32 param_2);
 u64 FUN_00357830(int param_1);
 void FUN_00357930(u32 param_1);
-u64 FUN_003579b0(u64 param_1);
+u32 FUN_003579b0(u32 param_1);
 void FUN_00357ba0(int param_1);
 void FUN_00357c10(int param_1);
 void FUN_00357c80(int param_1);
@@ -45668,7 +45668,7 @@ u32 FUN_0034bf10(int param_1)
 // FUN_0034BFC0 NONMATCHING
 
 
-u64 FUN_0034bfc0(u32 param_1)
+u32 FUN_0034bfc0(u32 param_1)
 
 
 
@@ -45676,7 +45676,7 @@ u64 FUN_0034bfc0(u32 param_1)
 
   u32 uVar1;
 
-  u64 uVar3;
+  u32 uVar3;
 
   u32 uVar4;
 
@@ -55774,7 +55774,7 @@ void FUN_00357930(u32 param_1)
 // FUN_003579B0 NONMATCHING
 
 
-u64 FUN_003579b0(u64 param_1)
+u32 FUN_003579b0(u32 param_1)
